backend/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==3.2.19
django-cors-headers==4.1.0
django-filter==23.2
django-redis==5.3.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.7.0
kombu==5.3.1
Markdown==3.4.3
prompt-toolkit==3.0.38
psycopg2-binary==2.8.6
python-dateutil==2.8.2
python-decouple==3.3
pytz==2023.3
redis==4.6.0
requests==2.31.0
six==1.16.0